app/requirements.txt
fastapi
uvicorn
pymysql
bcrypt
python-dotenv
google-cloud-storage
pillow
numpy
requests
streamlit